updated start.sh with skipping spark archive

+++ start.sh
@@ -1,8 +1,16 @@
 #!/bin/bash
-printf "\n-- Deleting spark archive --\n"
-rm spark/*.tgz
-printf "\n-- Downloading spark archive --\n"
-spark/get-spark.sh
+
+first_argument="skip"
+
+if [[ $1 = $first_argument ]]; then
+    printf "\n-- Skipping downloading spark archive --\n"
+else
+    printf "\n-- Deleting spark archive --\n"
+    rm spark/*.tgz
+    printf "\n-- Downloading spark archive --\n"
+    spark/get-spark.sh
+fi
+
 
 # Optional:
 # printf "\n-- Updating songs trainging set --\n"
